Optional로 데이터뿌리기

+++ src/main/java/com/apple/shop/ItemRepository.java
@@ -2,6 +2,8 @@ package com.apple.shop;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemRepository extends JpaRepository<Item, Integer> {
+import java.util.Optional;
+
+public interface ItemRepository extends JpaRepository<Item, Long> {
 
 }
